+++ gradle.properties
@@ -1,6 +1,6 @@
 org.gradle.jvmargs=-Xmx4096m
 
 group=org.polyfrost.oneconfig
-version=1.0.13
+version=1.0.13+SEARCH1
 
 ksp.incremental=false
